Change default dtype from fp64 to fp32 in tri routine

+++ chainerx_cc/chainerx/routines/creation.cc
@@ -196,7 +196,7 @@ Array Eye(int64_t n, absl::optional<int64_t> m, absl::optional<int64_t> k, absl:
         k = 0;
     }
     if (!dtype.has_value()) {
-        dtype = Dtype::kFloat64;
+        dtype = Dtype::kFloat32;
     }
     if (n < 0 || m < 0) {
         throw DimensionError{"Negative dimensions are not allowed"};
